+++ packages/zoom/src/ZoomOutButton.tsx
@@ -7,9 +7,8 @@
  */
 
 import * as React from 'react';
-import { LocalizationContext, MinimalButton, Position, Tooltip } from '@react-pdf-viewer/core';
+import { isMac, LocalizationContext, MinimalButton, Position, Tooltip } from '@react-pdf-viewer/core';
 
-import { isMac } from './utils/isMac';
 import { ZoomOutIcon } from './ZoomOutIcon';
 import type { RenderZoomOutProps } from './types/RenderZoomOutProps';