+++ packages/zoom/src/utils/isMac.ts
@@ -1,9 +1,0 @@
-/**
- * A React component to view a PDF document
- *
- * @see https://react-pdf-viewer.dev
- * @license https://react-pdf-viewer.dev/license
- * @copyright 2019-2021 Nguyen Huu Phuoc <me@phuoc.ng>
- */
-
-export const isMac = () => (typeof window !== 'undefined' ? /(Mac|iPhone|iPod|iPad)/i.test(navigator.platform) : false);